+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - Incorrect naming in README for generate_proto.py to generate_proto.sh
 - Changed README MIT license to Apache
 - Documentation structure updated: contents moved from `/documentation` to `/docs`.
+- Switched Mirror Node endpoints used by SDK to secure ones instead of deprecated insecure endpoints (shut down on Aug 20th, see [Hedera blogpost](https://hedera.com/blog/updated-deprecation-of-the-insecure-hedera-consensus-service-hcs-mirror-node-endpoints))
 
 ### Removed
 - Removed the old `/documentation` folder.